==> code/01_acquire_archive_audio.ipynb <==
## Step 1

We the Broadcastify Archive Tool (BArT) automates downloading audio archives from Broadcastify, "the world's largest source of Public Safety, Airline, Rail, and Marine live audio streams".

We sought out a large repository of emergency radio dispatches to train our audio-to-text-to-geolocation algorithm. Although Broadcastify's archive was the clear winner in terms of breadth and depth for this data, the lack of an API for downloading audio files meant that the process for acquiring the archive files was highly manual and time-consuming.

BArT solves that problem. Please see the documentation at https://github.com/ljhopkins2/BArT.

----
