==> code/04b_extract_geolocation_data_lat_long.ipynb <==
# ! pip install selenium

from bs4 import BeautifulSoup
import requests
import pickle
from time import sleep

from selenium import webdriver
from selenium.webdriver.support import wait
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC
from selenium.webdriver.common.by import By

To give our Google Speech to Text API and Mapping API context, we decided to scrape the street names from geographic.org for Fulton County in Atlanta, Georgia. We ran into some trouble using the BeautifulSoup API to webscrape this website page because it’s rendered in javascript, so we switched to Selenium, which handled the problem deftly. We saved our results and referenced them throughout the project. 

url = 'https://geographic.org/streetview/usa/ga/fulton/atlanta.html'
XPATH = "//a[@alt='Zoya Court']"

browser = webdriver.Chrome('../assets/chromedriver')
browser.get(url)

element = WebDriverWait(browser, 10).until(
                  EC.presence_of_element_located((By.XPATH, 
                                                  XPATH)))

soup = BeautifulSoup(browser.page_source, 'lxml')

streetlist = soup.find('ul')


anchors = streetlist.find_all('a')

streets = [anchor.text for anchor in anchors]

streets[:10]

with open('../data/Pickles/streets.pkl','wb') as pickle_out:
    pickle.dump(streets, pickle_out)
