==> code/05_plot_on_heatmap.ipynb <==
# Imports

#! pip install gmaps
#! pip install google

#! jupyter nbextension enable --py gmaps

import gmaps
import google
import requests
import gmaps.datasets

import pandas as pd
import numpy as np

gmaps.configure(api_key='<ENTER YOUR KEY HERE>')

# Get dataframes from csv
with open("../data/Pickles/streets.pkl", "rb") as fp:
    df_roads = pickle.load(fp)
df = pd.read_csv('../data/Data/Long_and_Lat_result_w_o_context/df_context.csv')
df_context = pd.read_csv('../data/Data/Long_and_Lat_result_w_o_context/df_context.csv')

# Mapping

This function accepts an address and a google api key to search google for the address given. It then saves the results as a JSON and parses through it to find the latitude and longitude of the address given, and saves those results. 

The Google Mapping API accepts a latitude and longitude to center the map around, and then maps exact road locations based on latitudes and longitudes passed into it. We reference the dataframe that contains the results from the Spacy NLP API for our latitude and longitude mapping coordinates. 

# SOURCE https://www.shanelynn.ie/batch-geocoding-in-python-with-google-geocoding-api/
# function that takes in as arguments a address and returns the latitude and longitude

def get_google_latlong(address, api_key=None, return_full_response=False):
    # Set up your Geocoding url
    geocode_url = "https://maps.googleapis.com/maps/api/geocode/json?address={}".format(address)
    if api_key is not None:
        geocode_url = geocode_url + "&key={}".format(api_key)
        
    # Ping google for the reuslts:
    results = requests.get(geocode_url)
    # Results will be in JSON format - convert to dict using requests functionality
    results = results.json()
    
    # if there's no results or an error, return empty results.
    if len(results['results']) == 0:
        output = {
            "latitude": None,
            "longitude": None
        }
    else:    
        answer = results['results'][0]
        output = {
            "latitude": answer.get('geometry').get('location').get('lat'),
            "longitude": answer.get('geometry').get('location').get('lng')
        }
        
    # Append some other details:    
    output['input_string'] = address
    output['number_of_results'] = len(results['results'])
    output['status'] = results.get('status')
    if return_full_response is True:
        output['response'] = results
    
    return output

# ENHANCED MODEL MAP
# Loop through the dataframe with the locations extracted and place a pin on the map for every one of them
# this gets the coordinated of each address from the api using the function above

# **Make sure to enable javascript version of google maps API for your account**

# we want to center the map on Fulton county so we need the coordinates to set the center
fulton_coordinates = (33.7576063, -84.3882602)
fig = gmaps.figure(center=fulton_coordinates,
             zoom_level=9, 
             map_type='TERRAIN', 
             layout={'width': '650px', 'height': '550px'})


# Loop through the results
for i in range(df.shape[0]):
    list_loc = df_context['location_extraction'].iloc[i].split(' , ')
    locations = []
    for i in range(len(list_loc)):
        x, y = (get_google_latlong(list_loc[i], api_key='<ENTER YOUR KEY HERE>')['latitude'], 
                get_google_latlong(list_loc[i], api_key='<ENTER YOUR KEY HERE>')['longitude'])
        if (x != None) & (y != None):
            locations.append((x, y))
    if len(locations) != 0:
        heatmap = gmaps.heatmap_layer(locations=locations)
        fig.add_layer(heatmap)
    
fig

>Code adopted from [NY General Assembly DSI radio-to-location repository](https://github.com/mchbmn/radio-to-location)
